inventory views
